requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
Deprecated==1.2.18
fonttools==4.59.0
frozenlist==1.7.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-tools==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
multidict==6.6.3
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pillow==11.3.0
propcache==0.3.2
protobuf==5.29.5
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wrapt==1.17.2
yandexcloud==0.351.0
yarl==1.20.1
ydb==3.21.6
waitress
python-dotenv==1.0.0
